% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{averagereturn}
\alias{averagereturn}
\title{Average (log) return of a time series of prices}
\usage{
averagereturn(backtest)
}
\arguments{
\item{backtest}{A \code{data frame} or a \code{matrix} containing time series
of prices. If there are several prices provided, only the first one will
be used.}
}
\value{
A \code{numeric} output giving the average annualized return as a
decimal number.
}
\description{
This function generates a time series of log-returns and
while requiring as an input a time series of prices. In the end, the return
scaled to an annual return by multplying it by 252.
}
\examples{
averagereturn(backtest = as.matrix(EuStockMarkets[, 1]))
averagereturn(backtest = as.data.frame(EuStockMarkets[, 2]))
}
