% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_construction.R
\name{equityanddebtpf}
\alias{equityanddebtpf}
\title{Generate a weighted portfolio of two assets}
\usage{
equityanddebtpf(equity, debt, equityaspercent)
}
\arguments{
\item{equity}{A \code{data frame} or \code{matrix} with a time series of
equity prices.}

\item{debt}{A \code{data frame} or \code{matrix} with a time series of
debt prices.}

\item{equityaspercent}{A \code{numeric input} indicating which proportion of
the PF should be invested in the first asset.}
}
\value{
A \code{data frame} containing the time series of the portfolio
performance.
}
\description{
The \code{function} takes two time series of asset prices as
input and estimates the portfolio evaluation based on the weight provided by
the user.
}
