% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_construction.R
\name{indexpf}
\alias{indexpf}
\title{Scale a time series of prices to a starting value of 100.}
\usage{
indexpf(data)
}
\arguments{
\item{data}{A \code{data frame} or \code{matrix} with a time series of
security prices}
}
\value{
A \code{data frame} with the evolution of the indexed PF
}
\description{
The \code{function} takes a \code{data frame} containing
security prices and scales this series to a starting value of 100.
}
\examples{
indexpf(EuStockMarkets[, 2:3])
}
