% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_construction.R
\name{minvarpf}
\alias{minvarpf}
\title{Function to get a PF, that applies minium variance calculations}
\usage{
minvarpf(data)
}
\arguments{
\item{data}{A \code{data frame} or \code{matrix} with a time series of
security prices}
}
\value{
A \code{data frame} with the evolution of the portfolio.
}
\description{
The goal of MVPFs is to find PF weights, that deliver the
smallest variance possible.
}
\examples{
minvarpf(EuStockMarkets[, 1:3])
}
