% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_construction.R
\name{optimpf}
\alias{optimpf}
\title{Calculate Sharpe ratio optimized PF weights}
\usage{
optimpf(data)
}
\arguments{
\item{data}{A \code{data frame} or \code{matrix} that contains the time
series of prices.}
}
\value{
A \code{data frame} with containing
time series of the portfolio performance.
}
\description{
The function can optimize a time series of prices in order to
get portfolio weights that deliver the highest possible Sharpe ratio. As this
function optimizes the Sharpe Ratio, which is calculated with a seperate
function in this package, the \code{sharpe function} needs to be loaded as
well.
}
\examples{
optimpf(EuStockMarkets[, 1:3])
}
