% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_construction.R
\name{riskparitypf}
\alias{riskparitypf}
\title{\code{Function} to get a PF, that follows the risk parity approach.}
\usage{
riskparitypf(equity, debt, commodity)
}
\arguments{
\item{equity}{A \code{data frame} or \code{matrix} with a time series of
security prices}

\item{debt}{A \code{data frame} or \code{matrix} with a time series of
security prices}

\item{commodity}{A \code{data frame} or \code{matrix} with a time series of
security prices}
}
\value{
A \code{data frame} with the evolution of the portfolio.
}
\description{
The goal of risk parity PFs is to find PF weights, such that the
risk contribution of each constutuent is the same for all PF constituents.
}
\examples{
riskparitypf(equity = EuStockMarkets[, 1], debt = EuStockMarkets[, 2], commodity = EuStockMarkets[, 3])
}
