% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{runDemo}
\alias{runDemo}
\title{Shiny app that runs the Gryffindor Robo Advisor.}
\usage{
runDemo()
}
\description{
With the Gryffindor Robo Advisor potential invetsors have the
unique possibility to find an in vestment possibility that is hand-selected
based on the investors needs and preferences. Beside the usual stuff required
to set up a portfolio (e.g. investment goals, capital available, ...) the
Gryffindor Robo Advisor applies insights from finance academia, to help the
investor asses her/his risk preference in a stable way.
}
\examples{
runDemo()
}
