% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{sharpe}
\alias{sharpe}
\title{Sharpe ratio calculation of a portfolio.}
\usage{
sharpe(expectedreturn, covmatrix, par)
}
\arguments{
\item{expectedreturn}{A \code{matrix} of expected returns.}

\item{covmatrix}{A \code{matrix} containing the covarinaces of the PF's
constituents, with their variance on the main diagonale}

\item{par}{A \code{matrix} containing the weights of the PF constituents.
All weights, except for one, have to be provided, as the last one is choosen,
such that the weights sum up to 100 percent.}
}
\value{
\code{Numeric vector} conraining the Sharpe ratio.
}
\description{
The function gives the Sharpe ratio of a PF, given its
constituents' expected returns, its covariances and weights.
}
\examples{
sharpe(expectedreturn = matrix(c(1, 2)),
covmatrix = matrix(c(0.004, 0.7, 0.7, 0.002), nrow = 2),
par = c(0.2))
}
