% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{yearlystd}
\alias{yearlystd}
\title{Yearly standard deviation of a time series of prices.}
\usage{
yearlystd(backtest)
}
\arguments{
\item{backtest}{A \code{data frame} or \code{matrix} containing one time
series of prices. If there are several prices provided, only the first one
will be used.}
}
\value{
A \code{numeric} output giving the yearly standard deviation
as decimalnumber
}
\description{
This function estimates the annualized standard deviation of a
time series of returns. It needs daily data as an input, as it scales the
daily standard deviation by the use of the square root of time rule to an
annual proxy.
}
\examples{
yearlystd(backtest = as.matrix(EuStockMarkets[, 1]))
yearlystd(backtest = as.data.frame(EuStockMarkets[, 2]))
}
